reorganized structure of LoginPage folder
