+++ src/components/LoginPage/LoginPageContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { REQUEST_AUTHORIZATION_ACTION } from 'components/LoginPage/LoginPagesActions';
-import { LoginPageComponent } from 'components/LoginPage/LoginPageComponents/LoginPageComponent';
+import { LoginPageComponent } from 'components/LoginPage/LoginPageComponent';
 import { action } from 'sagas';
 
 const mapStateToProps = ({ app: { isGoogleApiInitialized } }) => ({
