+++ src/components/LoginPage/Slide2.js
